+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   1. 支持 Rust, Haskell, OCaml 的 O2 优化
 
 ## 4.10.3
+
 - Fix:
   1. 复制代码按钮炸了